% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predictions.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Function to plot the Error Comparison}
\usage{
plot_predictions(dataIn)
}
\arguments{
\item{dataIn}{as input data in list format (returned by function prediction_errors())}
}
\value{
It returns the Error comparison for different methods
}
\description{
Function to plot the Error Comparison
}
\examples{
# aa <- prediction_errors()
# bb <- plot_predictions(aa)
# bb
}

