% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_append.R
\name{prediction_append}
\alias{prediction_append}
\title{To attach and compare new method to existing comparison study done with function 'impute_errors()'}
\usage{
prediction_append(existing_method, dataIn, nextVal, errorParameter, MethodPath,
  MethodName)
}
\arguments{
\item{existing_method}{as Error observations for different methods}

\item{dataIn}{as imput time series for testing}

\item{nextVal}{as an integer to decide number of values to predict}

\item{errorParameter}{as type of error calculation (RMSE, MAE or MAPE)}

\item{MethodPath}{as location of function for proposed imputation method}

\item{MethodName}{as name for function for proposed imputation method}
}
\value{
Returns error comparosin for imputation methods
}
\description{
To attach and compare new method to existing comparison study done with function 'impute_errors()'
}
\examples{
#Kindly, refer "Vignette" document
}

