% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_errors.R
\name{prediction_errors}
\alias{prediction_errors}
\title{Function working as testbench for comparison of Prediction algorithms}
\usage{
prediction_errors(dataIn, nextVal, errorParameter, MethodPath, MethodName)
}
\arguments{
\item{dataIn}{as input time series for testing}

\item{nextVal}{as an integer to decide number of values to predict}

\item{errorParameter}{as type of error calculation (RMSE, MAE or MAPE)}

\item{MethodPath}{as location of function for the proposed imputation method}

\item{MethodName}{as name for function for the proposed imputation method}
}
\value{
Returns error comparison for imputation methods
}
\description{
Function working as testbench for comparison of Prediction algorithms
}
\examples{
# aa <- prediction_errors(nextVal = 10)
# aa
}

