% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_ahead_forecast.R
\name{step_ahead_forecast}
\alias{step_ahead_forecast}
\title{Function to calculate the step ahead forecasting for a proposed Prediction method}
\usage{
step_ahead_forecast(dataIn, trainedData, MethodPath, errorParameter, stepSize)
}
\arguments{
\item{dataIn}{as input time series for testing}

\item{trainedData}{as partition point of input data `dataIn`}

\item{MethodPath}{as as location of function for the proposed imputation method}

\item{errorParameter}{as type of error calculation (RMSE, MAE or MAPE)}

\item{stepSize}{as interval at which step by step prediction will be done (Possible values are 1 & 2)}
}
\value{
returns the plot for one/two step ahead prediction along with error values decided by `errorParameter`
}
\description{
Function to calculate the step ahead forecasting for a proposed Prediction method
}

